% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{build_vectorizer}
\alias{build_vectorizer}
\title{build_vocab}
\usage{
build_vectorizer(text, id, max_features)
}
\arguments{
\item{text}{text vector}

\item{id}{id vector}
}
\value{
vectorizer
}
\description{
create vectorizer
}
