% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{check_list}
\alias{check_list}
\title{check_list}
\usage{
check_list(a, b)
}
\arguments{
\item{a}{a list of default parameters}

\item{b}{a list of input parameters}
}
\value{
out ...
}
\description{
check_list compares two lists and augments missing entries.
}
