% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{compile_keras_model}
\alias{compile_keras_model}
\title{compile_keras_model}
\usage{
compile_keras_model(container, target)
}
\arguments{
\item{container}{...}

\item{text}{...}
}
\value{
list(model = model, params = params, data = container$data)
}
\description{
Compile Keras graph
}
