% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_tidy_h2o_text.R, R/tidy_h2o.R
\name{fit_dnn}
\alias{fit_dnn}
\alias{fit_dnn}
\title{fit_dnn}
\usage{
fit_dnn(container)

fit_dnn(container)
}
\arguments{
\item{container}{...}

\item{container}{...}
}
\value{
model

model
}
\description{
fit Deep Neural Networks

fit Deep Neural Networks
}
