% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_tidy_h2o_text.R, R/tidy_h2o.R
\name{fit_gbm}
\alias{fit_gbm}
\alias{fit_gbm}
\title{fit_gbm}
\usage{
fit_gbm(container)

fit_gbm(container)
}
\arguments{
\item{container}{...}

\item{container}{...}
}
\value{
model

model
}
\description{
fit Gradient Boosting Machine

fit Gradient Boosting Machine
}
