% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_keras.R
\name{fit_glove}
\alias{fit_glove}
\title{fit_glove}
\usage{
fit_glove(container)
}
\arguments{
\item{container}{...}
}
\value{
list(model = model, params = params, data = container$data)
}
\description{
Fit Global average pooling operation for temporal data (Keras)
}
