% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_keras.R
\name{fit_lstm}
\alias{fit_lstm}
\title{fit_lstm}
\usage{
fit_lstm(container)
}
\arguments{
\item{container}{...}

\item{target}{...}
}
\value{
list(model = model, params = params, data = container$data)
}
\description{
Fit Long Short-Term Memory Network (Keras)
}
