% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_tidy_h2o_text.R, R/tidy_h2o.R
\name{fit_nb}
\alias{fit_nb}
\alias{fit_nb}
\title{fit_nb}
\usage{
fit_nb(container)

fit_nb(container)
}
\arguments{
\item{container}{...}

\item{container}{...}
}
\value{
model

model
}
\description{
fit Naive Bayes

fit Naive Bayes
}
