% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_tidy_h2o_text.R, R/tidy_h2o.R
\name{fit_xgboost}
\alias{fit_xgboost}
\alias{fit_xgboost}
\title{fit_xgboost}
\usage{
fit_xgboost(container)

fit_xgboost(container)
}
\arguments{
\item{container}{...}

\item{container}{...}
}
\value{
model

model
}
\description{
fit XGBoost

fit XGBoost
}
