% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{keras_predict}
\alias{keras_predict}
\title{keras_predict}
\usage{
keras_predict(model, test_input, index = 1)
}
\arguments{
\item{model}{...}

\item{test_input}{...}

\item{index}{correction}
}
\value{
preds
}
\description{
Convienience function for keras predictions
}
