% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mbo.R
\docType{data}
\name{metric}
\alias{metric}
\title{run_mbo}
\format{An object of class \code{character} of length 1.}
\usage{
metric
}
\arguments{
\item{data}{input}

\item{params}{params}

\item{const}{constants}

\item{n_init}{burn in iterations}

\item{n_main}{main iterations}
}
\value{
list(data = data, params = params)
}
\description{
tidy MBO
}
\keyword{datasets}
