% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mbo.R
\name{run_mbo}
\alias{run_mbo}
\title{run_mbo}
\usage{
run_mbo(data, params, target, text, n_init = 5, n_main = 30, name = "",
  metric, parallel)
}
\arguments{
\item{params}{params}

\item{target}{target variable}

\item{n_init}{burn in iterations}

\item{n_main}{main iterations}

\item{name}{the name}

\item{const}{constants}
}
\value{
list(data = data, params = params)
}
\description{
tidy MBO
}
