% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mbo.R
\name{run_mbo_steps}
\alias{run_mbo_steps}
\title{run_keras_steps}
\usage{
run_mbo_steps(params, data, target, text, metric, reconstruct = F)
}
\arguments{
\item{params}{params}

\item{target}{target variable}

\item{text}{text variable}
}
\value{
metric
}
\description{
run all tuning processes
}
