% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data_prep_old.R, R/data_prep.R
\name{split_data}
\alias{split_data}
\alias{split_data}
\title{split_data}
\usage{
split_data(data, p)

split_data(data, p)
}
\arguments{
\item{data}{input data}

\item{p}{propability of the train set}

\item{data}{input data}

\item{p}{propability of the train set}
}
\value{
out

out
}
\description{
split_data splits the data into train and test set

split_data splits the data into train and test set
}
