% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data_prep_old.R, R/data_prep.R
\name{text_to_matrix}
\alias{text_to_matrix}
\alias{text_to_matrix}
\title{h2o_text_to_matrix}
\usage{
text_to_matrix(container)

text_to_matrix(container)
}
\arguments{
\item{container}{...}

\item{container}{...}

\item{text}{...}
}
\value{
list(data = data, params = params)

list(data = data, params = params)
}
\description{
matrix tokenizer

matrix tokenizer
}
