% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data_prep_old.R, R/data_prep.R
\name{text_to_seq}
\alias{text_to_seq}
\alias{text_to_seq}
\title{text_to_seq}
\usage{
text_to_seq(container, text)

text_to_seq(container, text)
}
\arguments{
\item{container}{...}

\item{text}{...}

\item{container}{...}

\item{text}{...}
}
\value{
list(data = data, params = params)

list(data = data, params = params)
}
\description{
seq tokenizer

seq tokenizer
}
