% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R, R/data_prep_old.R
\name{build_dtm}
\alias{build_dtm}
\alias{build_dtm}
\title{build_dtm}
\usage{
build_dtm(vectorizer, text, id)

build_dtm(vectorizer, text, id)
}
\arguments{
\item{text}{...}

\item{container}{...}

\item{container}{...}

\item{text}{...}
}
\value{
list(data = data, params = params)

list(data = data, params = params)
}
\description{
build bag of words matrix

build bag of words matrix
}
