% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R, R/data_prep_old.R
\name{build_vectorizer}
\alias{build_vectorizer}
\alias{build_vectorizer}
\title{build_vocab}
\usage{
build_vectorizer(text, id, max_vocab = NULL, ngram = "unigram")

build_vectorizer(text, id, max_vocab = NULL, ngram = "unigram")
}
\arguments{
\item{text}{text vector}

\item{id}{id vector}

\item{text}{text vector}

\item{id}{id vector}
}
\value{
vectorizer

vectorizer
}
\description{
create vectorizer

create vectorizer
}
