% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{learn_h2o_model}
\alias{learn_h2o_model}
\title{learn_h2o_model}
\usage{
learn_h2o_model(container, target, reconstruct = F)
}
\arguments{
\item{container}{...}
}
\value{
list(metric = accuracy, params = params)
}
\description{
text pre-processing
}
