% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_keras.R
\name{learn_keras_model}
\alias{learn_keras_model}
\title{learn_keras_model}
\usage{
learn_keras_model(container, reconstruct)
}
\arguments{
\item{container}{...}
}
\value{
list(metric = accuracy, params = params)
}
\description{
text pre-processing
}
