% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{split_data}
\alias{split_data}
\title{split_data}
\usage{
split_data(data, p)
}
\arguments{
\item{data}{input data}

\item{p}{propability of the train set}
}
\value{
out
}
\description{
split_data splits the data into train and test set
}
