% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{text_to_matrix}
\alias{text_to_matrix}
\title{h2o_text_to_matrix}
\usage{
text_to_matrix(container)
}
\arguments{
\item{container}{...}
}
\value{
list(data = data, params = params)
}
\description{
matrix tokenizer
}
